requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
packaging==23.2
email-validator==2.1.0.post1
dnspython==2.4.2 